Lint and formated the code. Also made the karaoke page responsive.
