+++ lib/karaokium_web/live/admin/performance_live/edit.html.heex
@@ -5,5 +5,5 @@
   performance={@performance}
   title="Edit performance"
   action={:edit}
-  return_to={Routes.admin_karaoke_show_path(@socket, :show, @karaoke_id)}
+  return_to={Routes.admin_performance_show_path(@socket, :show, @karaoke_id, @performance)}
 />
