+++ lib/karaokium_web/templates/layout/_header.html.heex
@@ -4,7 +4,7 @@
       <img
         alt="Phoenix Framework Logo"
         src={Routes.static_path(@conn, "/karaokium/images/logos/cesium-DARK.svg")}
-        height="70"
+        height="50"
         style="margin: 0;"
       />
     <% end %>
